Virtual assistant using python/Aira/demo/requirements2.txt
datetime
SpeechRecognition
pyautogui
wikipedia
pywhatkit
pyobjc-framework-AppKit
requests
clipboard
pyjokes